Fix the "started on quadrant edge" problem
Say you start just to the left of the middle. If you swipe right and up:

Previous behavior: L' (because the gesture started on the left side)
New behavior:      R  (because the vertical part of the gesture started
    on the right side)

Similar change for U moves.

We'll see if this is more intuitive (and bug-free)!